+++ Plugins/Bugsnag/Source/Bugsnag/Private/Apple/ApplePlatformConfiguration.cpp
@@ -12,7 +12,8 @@
 
 static BSGTelemetryOptions GetTelemetryTypes(EBugsnagTelemetryTypes Value)
 {
-	return (EnumHasAllFlags(Value, EBugsnagTelemetryTypes::InternalErrors) ? BSGTelemetryInternalErrors : 0);
+	return (EnumHasAllFlags(Value, EBugsnagTelemetryTypes::InternalErrors) ? BSGTelemetryInternalErrors : 0) |
+		   (EnumHasAllFlags(Value, EBugsnagTelemetryTypes::Usage) ? BSGTelemetryUsage : 0);
 }
 
 static BSGThreadSendPolicy GetThreadSendPolicy(EBugsnagSendThreadsPolicy Policy)
@@ -107,6 +108,8 @@ BugsnagConfiguration* FApplePlatformConfiguration::Configuration(const TSharedRe
 	CocoaConfig.maxPersistedEvents = Configuration->GetMaxPersistedEvents();
 
 	CocoaConfig.maxPersistedSessions = Configuration->GetMaxPersistedSessions();
+
+	CocoaConfig.maxStringValueLength = Configuration->GetMaxStringValueLength();
 
 	CocoaConfig.persistUser = Configuration->GetPersistUser();