+++ Plugins/Bugsnag/Source/Bugsnag/Private/Apple/Specs/ApplePlatformConfiguration.spec.cpp
@@ -181,10 +181,13 @@ void FApplePlatformConfigurationSpec::Define()
 				{
 					TSharedRef<FBugsnagConfiguration> Configuration = MakeShared<FBugsnagConfiguration>(ApiKey);
 					BugsnagConfiguration* CocoaConfig = FApplePlatformConfiguration::Configuration(Configuration);
-					TEST_TRUE(CocoaConfig.telemetry == BSGTelemetryInternalErrors);
+					TEST_TRUE(CocoaConfig.telemetry == BSGTelemetryAll);
 					Configuration->SetTelemetry(EBugsnagTelemetryTypes::None);
 					CocoaConfig = FApplePlatformConfiguration::Configuration(Configuration);
 					TEST_TRUE(CocoaConfig.telemetry == 0);
+					Configuration->SetTelemetry(EBugsnagTelemetryTypes::Usage);
+					CocoaConfig = FApplePlatformConfiguration::Configuration(Configuration);
+					TEST_TRUE(CocoaConfig.telemetry == BSGTelemetryUsage);
 				});
 
 			It("ReleaseStage", [this]()